requirements.txt
Django==3.2
djangorestframework==3.12.4
drf-yasg==1.21.4